Split code generation
